+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - New function `uef_time:unix_time/0`
+- New function `uef_time:unix_time/1`
 
 ## [1.5.0] - 2019-05-12